feat: update base conversion
